+++ lib/aws_iot/adapters/tortoise/handler.ex
